Fix: new code if timed out
